+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 All notable changes to paopao-ce are documented in this file.
 ## 0.6.0+dev ([`dev`](https://github.com/rocboss/paopao-ce/tree/dev))
+TODO;
+
+## 0.5.1
 ### Added
 - User/Profile page add setting option when page's user is self.
 
-## 0.5.1
 ### Fixed
 - fixed update user metric incorrect when no user record.
+
+### Change
+- frontend: follow/unfollow option hint add username info in home/user/profile/message/collections page.
 
 ## 0.5.0
 ### Added